requirements.txt
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0 